add android version of screenwrapper
