+++ src/components/ScreenWrapper/BaseScreenWrapper.js
@@ -1,81 +1,25 @@
-import _ from 'underscore';
+import {KeyboardAvoidingView, View} from 'react-native';
 import React from 'react';
-import PropTypes from 'prop-types';
-import {View, KeyboardAvoidingView} from 'react-native';
 import {SafeAreaInsetsContext} from 'react-native-safe-area-context';
+import _ from 'underscore';
 import {withOnyx} from 'react-native-onyx';
-import styles from '../styles/styles';
-import * as StyleUtils from '../styles/StyleUtils';
-import HeaderGap from './HeaderGap';
-import KeyboardShortcutsModal from './KeyboardShortcutsModal';
-import KeyboardShortcut from '../libs/KeyboardShortcut';
-import onScreenTransitionEnd from '../libs/onScreenTransitionEnd';
-import Navigation from '../libs/Navigation/Navigation';
-import compose from '../libs/compose';
-import ONYXKEYS from '../ONYXKEYS';
-import CONST from '../CONST';
-import withNavigation from './withNavigation';
-import withWindowDimensions from './withWindowDimensions';
-import OfflineIndicator from './OfflineIndicator';
-import {withNetwork} from './OnyxProvider';
-import networkPropTypes from './networkPropTypes';
+import CONST from '../../CONST';
+import KeyboardShortcut from '../../libs/KeyboardShortcut';
+import Navigation from '../../libs/Navigation/Navigation';
+import onScreenTransitionEnd from '../../libs/onScreenTransitionEnd';
+import * as StyleUtils from '../../styles/StyleUtils';
+import styles from '../../styles/styles';
+import HeaderGap from '../HeaderGap';
+import KeyboardShortcutsModal from '../KeyboardShortcutsModal';
+import OfflineIndicator from '../OfflineIndicator';
+import compose from '../../libs/compose';
+import withNavigation from '../withNavigation';
+import withWindowDimensions from '../withWindowDimensions';
+import ONYXKEYS from '../../ONYXKEYS';
+import {withNetwork} from '../OnyxProvider';
+import {propTypes, defaultProps} from './propTypes';
 
-const propTypes = {
-    /** Array of additional styles to add */
-    style: PropTypes.arrayOf(PropTypes.object),
-
-    /** Returns a function as a child to pass insets to or a node to render without insets */
-    children: PropTypes.oneOfType([
-        PropTypes.node,
-        PropTypes.func,
-    ]).isRequired,
-
-    /** Whether to include padding bottom */
-    includePaddingBottom: PropTypes.bool,
-
-    /** Whether to include padding top */
-    includePaddingTop: PropTypes.bool,
-
-    // Called when navigated Screen's transition is finished.
-    onTransitionEnd: PropTypes.func,
-
-    /** Is the window width narrow, like on a mobile device */
-    isSmallScreenWidth: PropTypes.bool.isRequired,
-
-    /** The behavior to pass to the KeyboardAvoidingView, requires some trial and error depending on the layout/devices used.
-     *  Search 'switch(behavior)' in ./node_modules/react-native/Libraries/Components/Keyboard/KeyboardAvoidingView.js for more context */
-    keyboardAvoidingViewBehavior: PropTypes.oneOf(['padding', 'height', 'position']),
-
-    // react-navigation navigation object available to screen components
-    navigation: PropTypes.shape({
-        // Method to attach listener to Navigation state.
-        addListener: PropTypes.func.isRequired,
-    }),
-
-    /** Details about any modals being used */
-    modal: PropTypes.shape({
-        /** Indicates when an Alert modal is about to be visible */
-        willAlertModalBecomeVisible: PropTypes.bool,
-    }),
-
-    /** Information about the network */
-    network: networkPropTypes.isRequired,
-
-};
-
-const defaultProps = {
-    style: [],
-    includePaddingBottom: true,
-    includePaddingTop: true,
-    onTransitionEnd: () => {},
-    navigation: {
-        addListener: () => {},
-    },
-    modal: {},
-    keyboardAvoidingViewBehavior: 'padding',
-};
-
-class ScreenWrapper extends React.Component {
+class BaseScreenWrapper extends React.Component {
     constructor(props) {
         super(props);
 
@@ -158,8 +102,8 @@ class ScreenWrapper extends React.Component {
     }
 }
 
-ScreenWrapper.propTypes = propTypes;
-ScreenWrapper.defaultProps = defaultProps;
+BaseScreenWrapper.propTypes = propTypes;
+BaseScreenWrapper.defaultProps = defaultProps;
 
 export default compose(
     withNavigation,
@@ -170,4 +114,4 @@ export default compose(
         },
     }),
     withNetwork(),
-)(ScreenWrapper);
+)(BaseScreenWrapper);
