+++ src/components/ScreenWrapper/index.android.js
@@ -1,0 +1,17 @@
+import React from 'react';
+import BaseScreenWrapper from './BaseScreenWrapper';
+import {defaultProps, propTypes} from './propTypes';
+
+const ScreenWrapper = props => (
+    <BaseScreenWrapper
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        {...props}
+        keyboardAvoidingViewBehavior="height"
+    >
+        {props.children}
+    </BaseScreenWrapper>
+);
+ScreenWrapper.propTypes = propTypes;
+ScreenWrapper.defaultProps = defaultProps;
+
+export default ScreenWrapper;
